Ballon agregado, falta determinar aleatoriedad

+++ Practica SDL/HolaSDL/Vector2D.cpp
@@ -3,7 +3,7 @@
 
 
 void Vector2D::normalize() {
-	double mag = sqrt(pow(X, 2) + pow(Y, 2));
+	int mag = sqrt(pow(X, 2) + pow(Y, 2));
 	if (mag > 0.0) {
 		X = X / mag;
 		Y = Y / mag;
@@ -17,12 +17,12 @@ Vector2D Vector2D::operator+(const Vector2D& v) const {
 	return r;
 }
 
-Vector2D Vector2D::operator*(double d) const {
+Vector2D Vector2D::operator*(int d) const {
 	Vector2D r;
 	r.X = X * d;
 	r.X = Y * d;
 	return r;
 }
-double Vector2D::operator*(const Vector2D& d) const {
+int Vector2D::operator*(const Vector2D& d) const {
 	return d.X * X + d.Y * Y;
 }
